Detect that a React\Http\Request body has been completely read
React\Http\Server does not end a React\Http\Request until the
React\Http\Response is closed. If a listener wants to handle a Request's body
it needs to determine for itself that it has reached the end of that request's
body. Phluid\Request will now determine by means of comparing the Content-Length
header with the total number of bytes that it has seen.

If it's a chunked request the Content-Length header will be missing which means
we still need to rely on the React\Http\Request emitting an `end` event.